requirements.txt
streamlit==1.38.0
ultralytics==8.2.90
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
opencv-python-headless==4.10.0.84
numpy==1.26.0
Pillow==10.0.1
gdown==4.7.1
